Problem: our new site is in hugo
Solution: Start writing a port to styx.

First out is a minimal index.html, without layouts or anything, just
the content. And a bunch of scaffolding for filling in the rest bit
by bit.

The partials are in "markdown format", but it's just the html files
renamed. Rather than going about adding html as a passthrough file
type to styx, this is a decent quick workaround.

Any references to `{{ .Site.BaseURL }}` I simply replaced with `/`.

I found it confusing that what is more like static page content was
put among the more generic layouts and templates, so I separated it
into a `fractalide` theme for framing elements like header and footer
and generic page layouts, and a `fractalide-site` theme for elements
called from page content.

If this doesn't make sense to anyone else, `fractalide` and
`fractalide-site` are trivial to just smash together.

Build the new site with `nix-build new-site`.