Problem: documentation lacks header
The documentation page does not look like the others.

It comes off a bit brutal, as it jumps straight into the documentation
in a stark white, whereas the other pages have the nice gradient and
title fractal.

Solution: Use the `page.full` template, now that it exists.

Also add the nice footer gradient.

This change was the real reason for the `page.full` template in the
first place. :-)

Remove the interim `container` template.